refactor: group setting refactored
