+++ src/components/modals/group/GroupAddModal.tsx
@@ -17,26 +17,24 @@ export default function GroupAddModal() {
   const { checkedUsers } = useCheckedUsersStore();
   const [checkedGroups, setCheckedGroups] = useState<number[]>([]);
   const { mutate } = useAddGroupMember();
-  if (!user || !groups) return null;
 
   function groupSelectClickHandler(g: Group) {
-    if (checkedGroups.includes(g.groupId)) {
+    if (checkedGroups.includes(g.groupId))
       setCheckedGroups(checkedGroups.filter((groupId) => groupId !== g.groupId));
-    } else {
-      setCheckedGroups([...checkedGroups, g.groupId]);
-    }
+    else setCheckedGroups([...checkedGroups, g.groupId]);
   }
 
   function groupAddClickHandler() {
     checkedGroups.forEach((groupId) => mutate({ groupId, addMembers: [...checkedUsers] }));
   }
 
+  if (!user || !groups) return null;
   return (
     <Dialog>
       <DialogTrigger asChild>
         <Button
-          className="h-5 gap-2 rounded-full border-2
-          border-darkblue bg-white px-2 text-xs text-darkblue hover:bg-gray-200  md:h-8  md:px-3 lg:text-xl"
+          className="h-5 gap-2 rounded-full border-2 border-darkblue bg-white px-2 text-xs text-darkblue
+        hover:bg-gray-200 md:h-8 md:px-3 lg:text-xl"
         >
           다른 그룹에 추가
         </Button>
@@ -53,7 +51,7 @@ export default function GroupAddModal() {
                     checkedGroups.includes(g.groupId)
                       ? 'bg-darkblue text-white hover:bg-gray-500'
                       : 'border-2 border-darkblue bg-white text-darkblue hover:bg-gray-200'
-                  } gap-2 border-0 border-darkblue px-3 py-1  text-xl`}
+                  } gap-2 border-0 border-darkblue px-3 py-1 text-xl`}
                   onClick={() => groupSelectClickHandler(g)}
                 >
                   {g.groupName}
