+++ src/components/modals/group/setting/contents/index.ts
@@ -1,0 +1,2 @@
+export * from './DeleteContent';
+export * from './RenameContent';
