+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import Header from '@/components/header/Header';
 import MyProfileCard from '@/components/cards/MyProfileCard';
 import Groups from '@/components/group/Groups';
-import Footer from '@/components/Footer';
+import Footer from '@/components/utils/Footer';
 import ProfileSkeleton from '@/components/utils/ProfileSkeleton';
 import AgreementModal from '@/components/modals/AgreementModal';
 import Divider from '@/components/utils/Divider';
